backend/requirements.txt
click==8.0.4
dnspython==2.2.1
email-validator==1.1.3
flake8==4.0.1
Flask==2.0.3
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-WTF==1.0.0
gunicorn==20.0.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mccabe==0.6.1
mongoengine==0.24.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pymongo==4.0.2
python-decouple==3.6
six==1.16.0
Werkzeug==2.0.3
WTForms==3.0.1
